Hunt the Wumpus
PROJECT NAME : GAMERS Hunt the Wumpus is a simple two person game, originally created by Gregory Yob in 1972. It was created as a text-based game using BASIC. The game follows a hide-andseek format wherein a mysterious monster called the Wumpus lurks deep inside a networks of rooms. We have improvised on this basic idea, and planning to develop a graphical game featuring multiplayer functionality, which is both secure and robust. It includes a registration system which allows new users to register themselves on the database, and a login system that facilitates registered users to access the game system. Our game follows the concept of creating multiple game sessions, each of which has a capacity of 2 players. The players can start or join the existing game sessions. Once two registered players join the session, they are displayed with the game GUI along with the rules of the game and other instructions. The player who is successful in killing the Wumpus first wins the match. To make the game challenging for the players, we have also planned to integrate the classic Wumpus strategic placements of pits- a trench in which if the player falls, he dies; bats- which pick the player from his current position on the map and throw him randomly anywhere on the map and arrows- which the players can use to shoot at the angry and fearsome Wumpus. The win/loss records of each player will be updated in database after each match, which can be viewed whenever they login to the game system.